_M   .
M    configure.ac
_M   templates/module.spec.template
M    dispatch/BESCatalogEntry.cc
M    dispatch/Makefile.am
_M   dispatch/BESTokenizer.h
M    dispatch/BESCatalogUtils.cc
_M   dispatch/BESTokenizer.cc
M    dispatch/BESCatalogEntry.h
M    dispatch/unit-tests/catT.cc
M    dispatch/unit-tests/Makefile.am
_M   OSX_Resources/InstallationCheck.proto
_M   OSX_Resources/Info.plist.proto
_M   hello_world
_M   hello_world/bes-testsuite

+++ dispatch/BESCatalogEntry.cc
@@ -58,7 +58,7 @@ BESCatalogEntry::~BESCatalogEntry( )
 }
 
 void
-BESCatalogEntry::set_size( int size )
+BESCatalogEntry::set_size( off_t size )
 {
     ostringstream strm ;
     strm << size ;
